Treating the warning mentionned by Allan Cornet

+++ scilab/routines/scicos/extract_bit_16_LH.c
@@ -10,7 +10,7 @@ void extract_bit_16_LH(scicos_block *block,int flag)
    maxim=16;
    ref=0;
    for(i=0;i<maxim/2;i++)
-       {n=pow(2,i);
+       {n=(short)pow(2,i);
         ref=ref+n;}
    *y=(*u)&(ref);
 }
